infosubvenciones-api/Ingestion/requirements.txt
# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
alembic>=1.12.0

# Task Queue
celery>=5.3.0
redis>=5.0.0

# API Client & HTTP
requests>=2.31.0
urllib3>=2.0.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# PDF Processing
pymupdf>=1.23.0  # Primary PDF text extraction
# marker-pdf>=0.2.0  # Fallback for complex PDFs (optional - to be implemented)
# pypdf2>=3.0.0  # Alternative parser (optional)

# LLM & Embeddings
google-generativeai>=0.3.0  # Gemini API (LLM + embeddings)
# openai>=1.0.0  # OpenAI embeddings (optional - using Gemini instead)

# Utils
python-dotenv>=1.0.0
tenacity>=8.2.0  # Retry logic with exponential backoff
tqdm>=4.66.0  # Progress bars
loguru>=0.7.0  # Better logging

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
